revision of the oboInOwl handling
These changes are in response to:
https://github.com/UCDenver-ccp/datasource/issues/5

This revision of the previous fix for handling the invalid oboInOWL
namespace used by the OWL API OBO parser encapsulates the fix such that
it targets only the oboInOwl namespace. Calls to .equalsIgnoreCase have
been removed, replaced by a dedicated method for retrieving the
property URI from an OWLAnnotation class. Unit tests now test the
OntologyUtil.getNamespace() method as well.