Removed hard-coded file name for refseq release catalog file
The previous version of the RefSeq release catalog file parser used a
hard coded file name that needed to be changed periodically as the
release number for RefSeq is incremented at each release. The current
version now first retrieves the release number, then the correct file.
Fields for the workDirectory and clean flag were added to
FileRecordReader in order to facilitate this revision to the RefSeq
catalog parser.